soft_prompt/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soft-prompt-repro"
version = "0.1.0"
description = "Pragmatic reproduction scaffold for soft prompt tuning with frozen T5 models"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "datasets>=2.20.0",
  "transformers>=4.46.0",
  "torch>=2.5.0",
  "sentencepiece>=0.2.0",
  "protobuf>=4.25.0",
  "tqdm>=4.66.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.5.0",
  "matplotlib>=3.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
soft-prompt-repro = "soft_prompt_repro.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]